Extract NetworkScene
